Update ensureDir.js
Having the Archive flag set on a directory causes ensureDir() to recognize it exists, but not that it is a directory. This is because the attribute is set to 0x30 because both FILE_ATTRIBUTE_ARCHIVE (0x20) and FILE_ATTRIBUTE_DIRECTORY (0x10) are set. Changing === to & will fix this issue. A related issue could occur with FILE_ATTRIBUTE_HIDDEN (0x2), FILE_ATTRIBUTE_READONLY (0x1), FILE_ATTRIBUTE_SYSTEM (0x4), and potentially FILE_ATTRIBUTE_TEMPORARY (0x100).